♻️RequestHandelr가 너무 무거워지는것같아서 역할을 분리했습니다
